backend/requirements.txt
fastapi
sqlmodel
asyncpg
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
